+++ tests/test_async_tester.py
@@ -1,3 +1,8 @@
+import time
+from functools import partial
+
+import pytest
+from pytestqt import exceptions
 from qt_async_threads.pytest_plugin import AsyncTester
 
 
@@ -15,3 +20,20 @@ def test_start_and_wait(async_tester: AsyncTester) -> None:
 
     async_tester.start_and_wait(main())
     assert steps == ["main", "inner"]
+
+
+def test_start_and_wait_timeout(async_tester: AsyncTester) -> None:
+
+    async def main() -> None:
+        await async_tester.runner.run(partial(time.sleep, 2.0))
+
+    async_tester.start_and_wait(main())
+
+    # Test parameter.
+    with pytest.raises(exceptions.TimeoutError):
+        async_tester.start_and_wait(main(), timeout_s=1)
+
+    # Test default.
+    async_tester.timeout_s = 1
+    with pytest.raises(exceptions.TimeoutError):
+        async_tester.start_and_wait(main())